requirements.txt
MetaTrader5==5.0.5260
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
python-dotenv==1.0.1
loguru==0.7.2
joblib==1.4.2
